# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.createExaminationAppointmentByPatientRequest:
    properties:
      examination_schedule_id:
        type: integer
      patient_note:
        type: string
      payment_id:
        type: integer
    required:
    - examination_schedule_id
    - payment_id
    type: object
  api.createPatientRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
    required:
    - email
    - full_name
    - password
    - phone_number
    type: object
  api.loginUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  api.loginUserResponse:
    properties:
      access_token:
        type: string
      user_info:
        $ref: '#/definitions/api.userInfo'
    type: object
  api.userInfo:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      phone_number:
        type: string
      role:
        type: string
    type: object
  db.Booking:
    properties:
      appointment_date:
        type: string
      created_at:
        type: string
      id:
        type: integer
      patient_id:
        type: integer
      patient_note:
        type: string
      payment_id:
        type: integer
      payment_status:
        type: string
      status:
        type: string
      total_cost:
        type: integer
      type:
        type: string
    type: object
  db.GetExaminationAppointmentDetailsRow:
    properties:
      booking_id:
        type: integer
      booking_status:
        type: string
      dentist_name:
        type: string
      dentist_specialty:
        type: string
      end_time:
        type: string
      patient_note:
        type: string
      payment_status:
        type: string
      room_name:
        type: string
      start_time:
        type: string
      total_cost:
        type: integer
    type: object
  db.ListExaminationSchedulesByDateAndServiceCategoryRow:
    properties:
      dentist_name:
        type: string
      end_time:
        type: string
      room_name:
        type: string
      schedule_id:
        type: integer
      start_time:
        type: string
      type:
        type: string
    type: object
  db.Payment:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  db.Service:
    properties:
      category_id:
        type: integer
      cost:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      unit:
        type: string
      warranty_duration:
        type: string
    type: object
  db.ServiceCategory:
    properties:
      banner_url:
        type: string
      cost:
        type: integer
      created_at:
        type: string
      description:
        type: string
      icon_url:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: cvphuoc2014@gmail.com
    name: Châu Vĩnh Phước
  description: API documentation for the Dental Clinic application.
  title: Dental Clinic API
  version: 1.0.0
paths:
  /patients:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.userInfo'
        "400":
          description: Bad Request
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - accessToken: []
      summary: Lấy thông tin bệnh nhân
      tags:
      - users
  /patients/appointments/examination:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: List of examination bookings
          schema:
            items:
              $ref: '#/definitions/db.Booking'
            type: array
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - accessToken: []
      summary: Lấy tất cả danh sách lịch khám của bệnh nhân
      tags:
      - appointments
    post:
      consumes:
      - application/json
      parameters:
      - description: Examination Appointment Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createExaminationAppointmentByPatientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - accessToken: []
      summary: Cho phép bệnh nhân đặt lịch khám
      tags:
      - appointments
  /patients/appointments/examination/{id}:
    get:
      parameters:
      - description: Examination Appointment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.GetExaminationAppointmentDetailsRow'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      security:
      - accessToken: []
      summary: Lấy thông tin chi tiết của một lịch khám
      tags:
      - appointments
  /payment-methods:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Payment'
            type: array
        "500":
          description: Internal Server Error
      summary: Liệt kê tất cả phương thức thanh toán
      tags:
      - payments
  /schedules/examination:
    get:
      parameters:
      - description: Date in the format YYYY-MM-DD
        in: query
        name: date
        required: true
        type: string
      - description: Service Category ID
        in: query
        name: service_category_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.ListExaminationSchedulesByDateAndServiceCategoryRow'
            type: array
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Liệt kê tất cả lịch khám trong một ngày của một danh mục dịch vụ
      tags:
      - schedules
  /service-categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.ServiceCategory'
            type: array
        "500":
          description: Internal Server Error
      summary: Liệt kê tất cả danh mục dịch vụ hiện có
      tags:
      - services
  /service-categories/{slug}:
    get:
      parameters:
      - description: Category Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.ServiceCategory'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Lấy thông tin của một danh mục dịch vụ
      tags:
      - services
  /service-categories/{slug}/services:
    get:
      parameters:
      - description: Category Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Service'
            type: array
        "500":
          description: Internal Server Error
      summary: Liệt kê tất cả dịch vụ của một danh mục
      tags:
      - services
  /users:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create patient info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createPatientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
      summary: Tạo mới tài khoản bệnh nhân
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login user info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.loginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.loginUserResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Đăng nhập
      tags:
      - users
securityDefinitions:
  accessToken:
    description: JWT Authorization header using the Bearer scheme.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"